otpl: separate singularity JSON code
